# CourseZero/API Document/swagger.json
{"swagger":"2.0","info":{"version":"2019/2/19 下午 08:53:36","title":"API Reference"},"paths":{"/api/Avatar/UploadAvatar":{"post":{"tags":["Avatar"],"summary":"Keys are (first one MUST be auth_token):\r\nauth_token,\r\nfile,\r\n[file has to be less than 1MB]","operationId":"UploadAvatar","consumes":["multipart/form-data"],"produces":[],"parameters":[],"responses":{"200":{"description":"Success"}}}},"/api/Avatar/GetAvatarByUserid":{"post":{"tags":["Avatar"],"summary":"Return Avatars/default.png if userid is invalid","operationId":"GetAvatarByUserid","consumes":["application/json"],"produces":[],"parameters":[{"name":"avatarByUserid_Request","in":"body","description":"","required":false,"schema":{"$ref":"#/definitions/GetAvatarByUserid_Request"}}],"responses":{"200":{"description":"Success"}}}},"/api/ChangePassword":{"post":{"tags":["ChangePassword"],"operationId":"Post","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/Password_Change_Request"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Password_Change_Response"}}}}},"/api/ForgotPassword/Request_Change":{"post":{"tags":["ForgotPassword"],"operationId":"Request_Change","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/ForgotPassword_Change_Request"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/ForgotPassword_Change_Response"}}}}},"/api/ForgotPassword/Activate/{id}/{hash}":{"get":{"tags":["ForgotPassword"],"operationId":"Activate","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"hash","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"string"}}}}},"/api/General/GetAllCourses":{"post":{"tags":["General"],"summary":"Return a list of Course object on success. View class Course in CourseContext.cs for model info. Return \"status_code\": 1 when fail.","operationId":"GetAllCourses","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"authToken_Request","in":"body","description":"","required":false,"schema":{"$ref":"#/definitions/AuthToken_Request"}}],"responses":{"200":{"description":"Success"}}}},"/api/General/GetUserInfo":{"post":{"tags":["General"],"operationId":"GetUserInfo","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"authToken_Request","in":"body","required":false,"schema":{"$ref":"#/definitions/AuthToken_Request"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/GetUserInfo_Response"}}}}},"/api/Login":{"post":{"tags":["Login"],"operationId":"Post","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/Login_Request"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Login_Response"}}}}},"/api/Register":{"post":{"tags":["Register"],"operationId":"Post","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/Register_Request"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Register_Response"}}}}},"/api/Register/Reissue_Email":{"post":{"tags":["Register"],"summary":"Reissue an verification email","operationId":"Reissue_Email","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"request","in":"body","description":"","required":false,"schema":{"$ref":"#/definitions/Reissue_Email_Request"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Reissue_Email_Response"}}}}},"/api/Register/Verify_Email/{username}/{hash}":{"get":{"tags":["Register"],"operationId":"Verify_Email","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"username","in":"path","required":true,"type":"string"},{"name":"hash","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"string"}}}}},"/api/Session/Logout":{"post":{"tags":["Session"],"operationId":"Logout","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/Logout_Request"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Logout_Response"}}}}},"/api/Session/Get_All_Sessions":{"post":{"tags":["Session"],"operationId":"Get_All_Sessions","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/Get_All_Sessions_Request"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Get_All_Sessions_Response"}}}}},"/api/Session/Logout_Specific_Session":{"post":{"tags":["Session"],"operationId":"Logout_Specific_Session","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/Logout_Specific_Session_Request"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Logout_Specific_Session_Response"}}}}},"/api/Upload/UploadFile":{"post":{"tags":["Upload"],"summary":"Keys are (first one MUST be auth_token):\r\nauth_token,\r\nfile_name,\r\nfile_description,\r\nfile,\r\n[file has to be less than 100MB]","operationId":"UploadFile","consumes":["multipart/form-data"],"produces":[],"parameters":[],"responses":{"200":{"description":"Success"}}}}},"definitions":{"GetAvatarByUserid_Request":{"required":["auth_token","user_id"],"type":"object","properties":{"auth_token":{"maxLength":128,"minLength":128,"type":"string"},"user_id":{"format":"int32","description":"The avatar owner","type":"integer"}}},"Password_Change_Request":{"required":["auth_token","old_password","new_password"],"type":"object","properties":{"auth_token":{"maxLength":128,"minLength":128,"type":"string"},"old_password":{"maxLength":20,"minLength":5,"type":"string"},"new_password":{"maxLength":20,"minLength":5,"type":"string"}}},"Password_Change_Response":{"type":"object","properties":{"status_code":{"format":"int32","description":"<para>0 is success, 1, 2, 3 is fail</para>\r\n<para>1 fail is due to invalid auth token</para>\r\n<para>2 fail is due to invalid old password</para>\r\n<para>3 fail is due to invalid new password</para>\r\n<para>all current sessions will be logged out if success</para>","type":"integer"},"display_message":{"type":"string"}}},"ForgotPassword_Change_Request":{"required":["email"],"type":"object","properties":{"email":{"description":"A new password and an activation link will be sent to the user email","maxLength":27,"minLength":27,"type":"string"}}},"ForgotPassword_Change_Response":{"type":"object","properties":{"status_code":{"format":"int32","description":"<para>0 is success, 1 is fail</para>\r\n<para>all current sessions will be logged out if success</para>","type":"integer"},"display_message":{"type":"string"}}},"AuthToken_Request":{"required":["auth_token"],"type":"object","properties":{"auth_token":{"maxLength":128,"minLength":128,"type":"string"}}},"GetUserInfo_Response":{"type":"object","properties":{"status_code":{"format":"int32","description":"1 is fail, 0 is success","type":"integer"},"username":{"type":"string"}}},"Login_Request":{"required":["using_email","password","recaptcha_hash"],"type":"object","properties":{"using_email":{"description":"Should be TRUE if the user is trying to login with @link email. Otherwise, an username is expected.","type":"boolean"},"username":{"maxLength":20,"minLength":5,"type":"string"},"email":{"maxLength":27,"minLength":27,"type":"string"},"password":{"maxLength":20,"minLength":5,"type":"string"},"recaptcha_hash":{"type":"string"}}},"Login_Response":{"type":"object","properties":{"status_code":{"format":"int32","description":"0 is success, 1 is fail","type":"integer"},"display_message":{"type":"string"},"auth_token":{"description":"<para>A hash that serve as a credential for later purposes </para>\r\n<para>Should be stored in localstorage / sessionstorage (if the user does not check the \"remember me\" option)</para>","type":"string"},"username":{"type":"string"}}},"Register_Request":{"required":["username","password","email","recaptcha_hash"],"type":"object","properties":{"username":{"maxLength":20,"minLength":5,"type":"string"},"password":{"maxLength":20,"minLength":5,"type":"string"},"email":{"maxLength":27,"minLength":27,"type":"string"},"recaptcha_hash":{"type":"string"}}},"Register_Response":{"type":"object","properties":{"status_code":{"format":"int32","description":"0 is success, 1 is fail","type":"integer"},"display_message":{"type":"string"}}},"Reissue_Email_Request":{"required":["email"],"type":"object","properties":{"email":{"maxLength":27,"minLength":27,"type":"string"}}},"Reissue_Email_Response":{"type":"object","properties":{"status_code":{"format":"int32","description":"0 is success, 1 is fail","type":"integer"},"display_message":{"type":"string"}}},"Logout_Request":{"required":["auth_token"],"type":"object","properties":{"auth_token":{"maxLength":128,"minLength":128,"type":"string"}}},"Logout_Response":{"type":"object","properties":{"status_code":{"format":"int32","description":"<para>0 is success, 1 is fail</para>\r\n<para>fail may due to invalid auth token (user has already logged out)</para>","type":"integer"}}},"Get_All_Sessions_Request":{"required":["auth_token"],"type":"object","properties":{"auth_token":{"maxLength":128,"minLength":128,"type":"string"}}},"Get_All_Sessions_Response":{"type":"object","properties":{"status_code":{"format":"int32","description":"<para>0 is success, 1 is fail</para>\r\n<para>fail is due to invalid auth token (user has already logged out)</para>","type":"integer"},"sessions":{"description":"Return a list of session information.","uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/AuthToken"}}}},"AuthToken":{"type":"object","properties":{"token":{"type":"string"},"userID":{"format":"int32","type":"integer"},"last_access_IP":{"type":"string"},"last_access_Location":{"type":"string"},"last_access_Time":{"format":"date-time","type":"string"},"last_access_Device":{"type":"string"},"last_access_Browser":{"type":"string"}}},"Logout_Specific_Session_Request":{"required":["auth_token","token_to_be_removed"],"type":"object","properties":{"auth_token":{"maxLength":128,"minLength":128,"type":"string"},"token_to_be_removed":{"description":"<para>the token that the user want to log it out</para>\r\n<para>should prevent the user logout its own session\r\n            this part is not checked on server side</para>","maxLength":128,"minLength":128,"type":"string"}}},"Logout_Specific_Session_Response":{"type":"object","properties":{"status_code":{"format":"int32","description":"<para>0 is success, 1, 2 is fail</para>\r\n<para>1 fail is due to invalid auth token (user has already logged out)</para>\r\n<para>2 fail is due to given session does not exist or not valid</para>","type":"integer"}}}}}